requirements.txt
requests
PyYAML
psycopg2
